requirements.txt
jieba~=0.42.1
requests==2.28.2
Pillow==9.4.0
pytz==2022.7.1
prettytable==3.6.0
nb-cli==1.0.5
selenium==4.8.2
bs4~=0.0.1
beautifulsoup4==4.11.2
matplotlib==3.7.0
pymongo==4.3.3